* commit 'c6b5f0018c35dc2d8d07f792cf8bbafa37098995':
  댓글 수정
  채팅 진행 중
  채팅 진행 중
  키워드 - 회원가입(user service)에 familyname, givenname 추가
  dto 수정
  키워드 - CorsFilter- authorization추가, post- givenName추가
  좋아요 수정
  토큰 받기

# Conflicts:
#	facebook/src/main/java/com/clone/facebook/dto/PostResponseDto.java
#	facebook/src/main/java/com/clone/facebook/filter/CorsFilter.java

+++ facebook/src/main/java/com/clone/facebook/service/CommentService.java
@@ -1,42 +1,74 @@
 package com.clone.facebook.service;
 
-import com.clone.facebook.dto.CommentDto;
+import com.auth0.jwt.interfaces.DecodedJWT;
+import com.clone.facebook.dto.CommentRequestDto;
+
+
+
 import com.clone.facebook.models.Comment;
+
+import com.clone.facebook.models.User;
 import com.clone.facebook.repository.CommentRepository;
+import com.clone.facebook.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+
 import java.util.List;
+
+import static com.auth0.jwt.JWT.decode;
 
 @RequiredArgsConstructor
 @Service
 public class CommentService {
     private final CommentRepository commentRepository;
+    private final UserRepository userRepository;
+    private User getUser(String authorization){
+        String token = authorization.substring(7);
+        DecodedJWT decodeToken = decode(token);
+        User user = userRepository.findById(Long.parseLong(decodeToken.getClaim("id").toString()))
+                .orElseThrow(() -> new NullPointerException("ID DOES NOT EXIST"));
+    return user;
+
+    }
 
     @Transactional
     public List<Comment> getComment(Long contentsId){
-        return commentRepository.findAllByContentsId(contentsId);
+        List<Comment> comments = commentRepository.findAllByContentsId(contentsId);
+        return comments;
     }
 
     @Transactional
-    public Long postComment(Long contentsId, CommentDto commentDto,Long userId){
-        Comment comment = new Comment(contentsId,commentDto,userId);
+    public Long postComment(Long contentsId, CommentRequestDto commentDto, String authorization){
+
+        Comment comment = new Comment(contentsId,commentDto,getUser(authorization));
         commentRepository.save(comment);
+
         return comment.getId();
     }
 
     @Transactional
-    public Long updateComment(Long contentsId,Long id,CommentDto commentDto,Long userId){
-        Comment comment = commentRepository.findByIdAndContentsId(id,contentsId);
-        comment.update(contentsId,commentDto,userId);
-        return comment.getId();
+    public Long updateComment(Long contentsId, Long id, CommentRequestDto commentDto, String authorization){
+        if(commentRepository.findByIdAndContentsId(id,contentsId).getUserId().equals(getUser(authorization).getId())){
+            Comment comment = commentRepository.findByIdAndContentsId(id,contentsId);
+            comment.update(contentsId,commentDto,getUser(authorization));
+
+            return comment.getId();
+        }else {
+            throw new IllegalArgumentException("회원정보가 일치하지 않습니다");
+        }
+
     }
 
     @Transactional
-    public Long deleteComment(Long contentsId,Long id){
-        Comment comment = commentRepository.findByIdAndContentsId(id,contentsId);
-        commentRepository.delete(comment);
-        return comment.getId();
+    public Long deleteComment(Long contentsId,Long id, String authorization){
+        if(commentRepository.findByIdAndContentsId(id,contentsId).getUserId().equals(getUser(authorization).getId())){
+            Comment comment = commentRepository.findByIdAndContentsId(id,contentsId);
+            commentRepository.delete(comment);
+            return comment.getId();
+        }else {
+            throw new IllegalArgumentException("회원정보가 일치하지 않습니다");
+        }
     }
 }
